+++ src/services/PaymentCurrencies.php
@@ -10,6 +10,7 @@ namespace craft\commerce\services;
 use Craft;
 use craft\commerce\db\Table;
 use craft\commerce\errors\CurrencyException;
+use craft\commerce\helpers\Currency as CurrencyHelper;
 use craft\commerce\models\PaymentCurrency;
 use craft\commerce\Plugin;
 use craft\commerce\records\Order;
@@ -19,10 +20,6 @@ use craft\helpers\ArrayHelper;
 use yii\base\Component;
 use yii\base\Exception;
 use yii\base\InvalidConfigException;
-use Money\Converter;
-use Money\Currency;
-use Money\Exchange\FixedExchange;
-use Money\Exchange\ReversedCurrenciesExchange;
 
 /**
  * Payment currency service.
@@ -44,21 +41,6 @@ class PaymentCurrencies extends Component
      * @var PaymentCurrency[]
      */
     private $_allCurrenciesById;
-
-    /**
-     * @var
-     */
-    private $_exchange;
-
-    /**
-     *
-     */
-    public function init()
-    {
-        $this->_exchange = new ReversedCurrenciesExchange(new FixedExchange([
-            $this->getPrimaryPaymentCurrencyIso() => ArrayHelper::map($this->getNonPrimaryPaymentCurrencies(), 'iso', 'rate')
-        ]));
-    }
 
     /**
      * Get payment currency by its ID.
@@ -195,10 +177,11 @@ class PaymentCurrencies extends Component
      * @param float $amount
      * @param string $fromCurrency
      * @param string $toCurrency
+     * @param bool $round
      * @return float
      * @throws CurrencyException if currency not found by its ISO code
      */
-    public function convertCurrency(float $amount, string $fromCurrency, string $toCurrency): float
+    public function convertCurrency(float $amount, string $fromCurrency, string $toCurrency, $round = false): float
     {
         $fromCurrency = $this->getPaymentCurrencyByIso($fromCurrency);
         $toCurrency = $this->getPaymentCurrencyByIso($toCurrency);
@@ -213,10 +196,16 @@ class PaymentCurrencies extends Component
 
         if ($this->getPrimaryPaymentCurrency()->iso != $fromCurrency) {
             // now the amount is in the primary currency
-            $amount = $amount / $fromCurrency->rate;
+            $amount /= $fromCurrency->rate;
         }
 
-        return $amount * $toCurrency->rate;
+        $result = $amount * $toCurrency->rate;
+
+        if ($round) {
+            return CurrencyHelper::round($result, $toCurrency);
+        }
+
+        return $result;
     }
 
 
@@ -324,6 +313,8 @@ class PaymentCurrencies extends Component
                 'iso',
                 'primary',
                 'rate',
+                'dateCreated',
+                'dateUpdated',
             ])
             ->from([Table::PAYMENTCURRENCIES]);
     }